Create README.md
Together, let's build a community engaged in fostering the heritage entrepreneurial ecosystem. The DHP Platform is designed as an online-based initiative (a one-stop shop) aimed at promoting, preserving, and sharing Africa's cultural heritage with both the African community and international audiences. The platform would serve as an informative hub, educational platform, and promotional tool for the African cultural scene. It will be accessible to everyone from the public to private sectors, including heritage practitioners, professionals, tourists, students, researchers, and the general public. It will be accessible to everyone from the public to private sectors, including heritage practitioners, professionals, tourists, students, researchers, and the general public.